[refactoring] Notnull한 컬럼 원시타입으로 수정

+++ gg-pingpong-api/src/test/java/gg/party/api/admin/penalty/PartyPenaltyControllerTest.java
@@ -148,7 +148,7 @@ public class PartyPenaltyControllerTest {
 
 			assertEquals(penaltyDto.getPenaltyType(), updatedPenalty.getPenaltyType());
 			assertEquals(penaltyDto.getMessage(), updatedPenalty.getMessage());
-			assertEquals(penaltyDto.getPenaltyTime(), updatedPenalty.getPenaltyTime().intValue());
+			assertEquals(penaltyDto.getPenaltyTime(), updatedPenalty.getPenaltyTime());
 		}
 
 		@Test
